:recycle: #250 - refactor: merge format user implementations
