+++ frontend/src/lib/auth/loaders.ts
@@ -1,9 +1,9 @@
 import { LoaderFunction } from "@remix-run/router/utils";
 import { redirect } from "react-router-dom";
 
-import { formatUser } from "../../pages/destructionlist/utils";
 import { whoAmI } from "../api/auth";
 import { DestructionList } from "../api/destructionLists";
+import { formatUser } from "../format/user";
 import {
   canReviewDestructionList,
   canStartDestructionList,
