+++ frontend/src/pages/destructionlist/utils.ts
@@ -23,20 +23,3 @@ export async function updateSelectedZaken(
         attributeData.length ? (attributeData as unknown as Zaak[]) : zaken,
       );
 }
-
-/**
- * Returns the correct format for a user.
- * @param user
- * @param showRole
- */
-export function formatUser(user: User, showRole = false) {
-  const displayName =
-    user.firstName && user.lastName
-      ? `${user.firstName} ${user.lastName} (${user.username})`
-      : user.username;
-
-  if (showRole && user.role.name) {
-    return `${displayName} (${user.role.name})`;
-  }
-  return displayName;
-}
