+++ frontend/src/pages/landing/Landing.tsx
@@ -23,13 +23,13 @@ import {
   canViewDestructionList,
 } from "../../lib/auth/permissions";
 import { timeAgo } from "../../lib/format/date";
+import { formatUser } from "../../lib/format/user";
 import {
   PROCESSING_STATUS_ICON_MAPPING,
   PROCESSING_STATUS_LEVEL_MAPPING,
   PROCESSING_STATUS_MAPPING,
   STATUS_MAPPING,
 } from "../constants";
-import { formatUser } from "../destructionlist/utils";
 import "./Landing.css";
 import { LandingContext } from "./Landing.loader";
 
@@ -145,7 +145,7 @@ export const Landing = () => {
 
       const footer = (
         <P muted size="xs">
-          {formatUser(currentAssignee, true)}
+          {formatUser(currentAssignee, { showRole: true })}
           {otherAssignees.length && (
             <strong className="LandingPage__assignees-count">
               {" "}
@@ -165,7 +165,7 @@ export const Landing = () => {
         assignees: otherAssignees.length ? (
           <Tooltip
             content={otherAssignees
-              .map((a) => formatUser(a.user, true))
+              .map((a) => formatUser(a.user, { showRole: true }))
               .join(", ")}
             placement="bottom"
           >
